Bug fix for RecordClass decoding: o  Fixed improper cast causing RecordClass values to truncate. o  Test case for above.
git-svn-id: https://svn.apache.org/repos/asf/directory/apacheds/branches/bigbang@604761 13f79535-47bb-0310-9956-ffa450edef68

+++ protocol-dns/src/main/java/org/apache/directory/server/dns/io/decoder/DnsMessageDecoder.java
@@ -163,7 +163,7 @@ public class DnsMessageDecoder
             String domainName = getDomainName( byteBuffer );
 
             RecordType recordType = RecordType.convert( byteBuffer.getShort() );
-            RecordClass recordClass = RecordClass.convert( ( byte ) byteBuffer.getShort() );
+            RecordClass recordClass = RecordClass.convert( byteBuffer.getShort() );
 
             questions.add( new QuestionRecord( domainName, recordType, recordClass ) );
         }
